# Optimize WriteString method for primitive objects