Replace fmt.Sprintf with strings.Builder and strconv conversions.
Resolves #348 

Benchmarks (ran on development computer):

Before
```
BenchmarkPdfObjectIntegerWriteString-8      	10000000	       106 ns/op
BenchmarkPdfObjectFloatWriteString-8        	 5000000	       257 ns/op
BenchmarkPdfObjectStringWriteString-8       	   10000	    113057 ns/op
BenchmarkPdfObjectNameWriteString-8         	  500000	      3190 ns/op
BenchmarkPdfObjectArrayWriteString-8        	      20	 100750574 ns/op
BenchmarkPdfObjectDictionaryWriteString-8   	     200	   6185730 ns/op
BenchmarkPdfObjectReferenceWriteString-8    	10000000	       162 ns/op
BenchmarkPdfIndirectObjectWriteString-8     	10000000	       109 ns/op
BenchmarkPdfObjectStreamWriteString-8       	10000000	       110 ns/op
BenchmarkPdfObjectStreamsWriteString-8      	20000000	       114 ns/op
```

After
```
BenchmarkPdfObjectIntegerWriteString-8      	50000000	        31.7 ns/op
BenchmarkPdfObjectFloatWriteString-8        	20000000	        95.1 ns/op
BenchmarkPdfObjectStringWriteString-8       	   10000	      110947 ns/op
BenchmarkPdfObjectNameWriteString-8         	  500000	        3181 ns/op
BenchmarkPdfObjectArrayWriteString-8        	    3000	      435748 ns/op
BenchmarkPdfObjectDictionaryWriteString-8   	     300	     4658395 ns/op
BenchmarkPdfObjectReferenceWriteString-8    	20000000	         153 ns/op
BenchmarkPdfIndirectObjectWriteString-8     	20000000	        82.8 ns/op
BenchmarkPdfObjectStreamWriteString-8       	20000000	        84.6 ns/op
BenchmarkPdfObjectStreamsWriteString-8      	20000000	        85.2 ns/op
```